printer_agent/requirements.txt
python-socketio[client]
python-escpos

